Move out utils and add tests
