+++ src/main/scala/ca/valencik/Utils.scala
@@ -14,4 +14,20 @@ object Utils {
     putStrLn("Elapsed time: " + (t1 - t0) + "ns")
     result
   }
+
+  def infiniteStream[A](xs: Iterable[A]): Stream[A] = {
+    lazy val ss: Stream[A] = xs.toStream #::: ss
+    ss
+  }
+
+  def firstDuplicate[A](xs: Iterable[A]): A = {
+    var these: Iterable[A] = xs
+    var seen: Set[A]       = Set()
+    while (!seen(these.head)) {
+      seen += these.head
+      these = these.tail
+    }
+    these.head
+  }
+
 }
